Added support for LIAR and FNN datasets, as well as added lemmetization
For the liar dataset, anything labeled as "true" or "mostly-true" was considered as true, and all others to be fake. Text and titles were combined into one string and the CountVectorizer was modified to take in lemmetization to tokenize. Added support for converting numbers like 100,000 to 100000. Moved weekly tasks to s separate txt file.