Updated preprocessing- stopwords lemmetization and train/test option
All preprocessing functions use a lemmetized list of stopwords now so that we don't get random words like "ha" and "wa" and "s" that weren't getting filtered out. Also added a boolean variable (default is false) to the Liar and FNN preprocessing functions to indicate if you want test (false) or training (true) data to be used. This increases the startup time since both datasets are held in memory, will want to modify this later.