Fixed some preprocessing things, simple NN added
Changed the preprocessing functions so that there is a separate function to just get the preprocessed text when we are testing and to get the document term matrix and vectorizer when training. Also added a training and testing data split for Coronavirus dataset. Finally, created a simple NN based off of our assignment 2 code and this article
https://medium.com/analytics-vidhya/part-1-sentiment-analysis-in-pytorch-82b35edb40b8